maj jeudi 13h

+++ src/Driver-SIMU/AsservDriver.hpp
@@ -122,7 +122,7 @@ public:
      float odo_GetY_mm();
      float odo_GetTheta_Rad();		// angle in radian
      float odo_GetTheta_Degree();		// angle in degrees*/
-    void odo_SetPosition(double x_m, double y_m, double angle_rad);
+    void odo_SetPosition(float x_m, float y_m, float angle_rad);
     RobotPosition odo_GetPosition();
     int path_GetLastCommandStatus();
     void path_InterruptTrajectory();
